+++ bin/configure.js
@@ -351,11 +351,11 @@
 
 				require(_path.resolve(_ROOT, './libraries/lychee/source/core/lychee.js'));
 				require(_path.resolve(_ROOT, './libraries/lychee/source/core/Asset.js'));
+				require(_path.resolve(_ROOT, './libraries/lychee/source/core/Package.js'));
 				require(_path.resolve(_ROOT, './libraries/lychee/source/core/Definition.js'));
 				require(_path.resolve(_ROOT, './libraries/lychee/source/core/Environment.js'));
-				require(_path.resolve(_ROOT, './libraries/lychee/source/core/Package.js'));
-				require(_path.resolve(_ROOT, './libraries/lychee/source/core/Simulation.js'));
 				require(_path.resolve(_ROOT, './libraries/lychee/source/core/Specification.js'));
+				require(_path.resolve(_ROOT, './libraries/lychee/source/core/Simulation.js'));
 				require(_path.resolve(_ROOT, './libraries/lychee/source/platform/node/features.js'));
 				require(_path.resolve(_ROOT, './libraries/lychee/source/platform/node/bootstrap.js'));