+++ libraries/breeder/bin/breeder.js
@@ -118,11 +118,11 @@ const _bootup = function(settings) {
 		sandbox:  settings.debug === true ? false : true,
 		build:    'breeder.Main',
 		timeout:  5000,
-		packages: [
-			new lychee.Package('lychee',     '/libraries/lychee/lychee.pkg'),
-			new lychee.Package('fertilizer', '/libraries/fertilizer/lychee.pkg'),
-			new lychee.Package('breeder',    '/libraries/breeder/lychee.pkg')
-		],
+		packages: {
+			'lychee':     '/libraries/lychee/lychee.pkg',
+			'fertilizer': '/libraries/fertilizer/lychee.pkg',
+			'breeder':    '/libraries/breeder/lychee.pkg'
+		},
 		tags:     {
 			platform: [ 'node' ]
 		}