+++ libraries/strainer/bin/strainer-fixer.js
@@ -65,10 +65,10 @@ const _bootup = function(settings) {
 		sandbox:  true,
 		build:    'strainer.Fixer',
 		timeout:  5000,
-		packages: [
-			new lychee.Package('lychee',   '/libraries/lychee/lychee.pkg'),
-			new lychee.Package('strainer', '/libraries/strainer/lychee.pkg')
-		],
+		packages: {
+			'lychee':   '/libraries/lychee/lychee.pkg',
+			'strainer': '/libraries/strainer/lychee.pkg'
+		},
 		tags:     {
 			platform: [ 'node' ]
 		}